+++ native/c/test/zowex.ds.test.cpp
@@ -830,6 +830,34 @@ void zowex_ds_tests()
                              Expect(response).ToContain("ABC2");
                              Expect(response).Not().ToContain("XYZ1");
                            });
+                        it("should return ISPF statistics attributes when listing members with attributes",
+                           [&]() -> void
+                           {
+                             string ds = get_random_ds();
+                             _ds.push_back(ds);
+
+                             _create_ds(ds, "--dsorg PO --dirblk 2");
+
+                             string response;
+
+                             execute_command_with_output(
+                                 zowex_command + " data-set create-member '" + ds + "(STAT1)'",
+                                 response);
+
+                             execute_command_with_output(
+                                 "echo \"TEST\" | " + zowex_command + " data-set write '" + ds + "(STAT1)'",
+                                 response);
+                             string command = zowex_command + " data-set lm " + ds + " --attributes";
+
+                             int rc = execute_command_with_output(command, response);
+
+                             ExpectWithContext(rc, response).ToBe(0);
+
+                             Expect(response).ToContain("STAT1");
+                             Expect(response).ToContain("/");
+                             Expect(response).ToContain(":");
+                             Expect(response).ToContain("N");
+                           });
                       });
              // TODO: https://github.com/zowe/zowe-native-proto/issues/380
              xdescribe("restore",